+++ modules/avm/datadog.tf
@@ -1,6 +1,6 @@
 module "datadog" {
   count                 = try(var.datadog.enable_integration, false) == true ? 1 : 0
-  source                = "github.com/schubergphilis/terraform-aws-mcaf-datadog?ref=v0.3.3"
+  source                = "github.com/schubergphilis/terraform-aws-mcaf-datadog?ref=v0.3.4"
   providers             = { aws = aws.managed_by_inception }
   api_key               = try(var.datadog.api_key, null)
   install_log_forwarder = try(var.datadog.install_log_forwarder, false)