+++ modules/avm/variables.tf
@@ -58,6 +58,7 @@ variable "create_workspace" {
 variable "datadog" {
   type = object({
     api_key               = string
+    datadog_tags          = list(string)
     enable_integration    = bool
     install_log_forwarder = bool
     site_url              = string